llamastack-practices/requirements.txt
llama-stack
llama-stack-client
python-dotenv
yfinance

uuid

datasets